+++ library/psr11/ContainerException.php
@@ -1,0 +1,10 @@
+<?php declare(strict_types=1);
+
+namespace suda\psr11;
+
+use Exception;
+use Psr\Container\ContainerExceptionInterface;
+
+class ContainerException extends Exception implements ContainerExceptionInterface
+{
+}