+++ library/psr11/NotFoundException.php
@@ -1,0 +1,10 @@
+<?php declare(strict_types=1);
+
+namespace suda\psr11;
+
+use Exception;
+use Psr\Container\NotFoundExceptionInterface;
+
+class NotFoundException extends Exception implements NotFoundExceptionInterface
+{
+}